fix: move rateLimiter to utils
