+++ app.js
@@ -6,7 +6,7 @@ const helmet = require('helmet');
 const cors = require('cors');
 const bodyParser = require('body-parser');
 const { errors } = require('celebrate');
-const { limiter } = require('./middlewares/rateLimiter');
+const { limiter } = require('./utils/rateLimiter');
 const { requestLogger, errorLogger } = require('./middlewares/logger');
 const { errorHandler } = require('./utils/errorHandler');
 const { CONFIG } = require('./config');
